+++ src/index.js
@@ -1,14 +1,8 @@
 import { trial } from './foa';
 import {
-  printCsvBestPositionPerTrial,
-  printCsvFoodSwarmCoordinatesPerSingleTrial,
-  printCsvFruitFlyCoordinatesPerSingleTrial,
-  printCsvSmellConcentrationPerSingleTrial,
+  printCsvDeltaPerSingleTrial 
 } from './print';
 
 const trialResults = trial();
 
-printCsvFruitFlyCoordinatesPerSingleTrial(trialResults);
-printCsvFoodSwarmCoordinatesPerSingleTrial(trialResults);
-printCsvSmellConcentrationPerSingleTrial(trialResults);
-printCsvBestPositionPerTrial(trialResults);
+printCsvDeltaPerSingleTrial(trialResults);