+++ src/swarm.js
@@ -17,7 +17,22 @@ export class Swarm {
   }
 
   vision(bestPosition) {
+    const delta = this._deriveDelta(bestPosition);
+    
     const { coordinates } = bestPosition;
     this._coordinates = coordinates;
+
+    return delta;
+  }
+
+  _deriveDelta(bestPosition) {
+    const { coordinates } = bestPosition;
+    const { x: xNext, y: yNext } = coordinates;
+    const { x: xCurrent, y: yCurrent } = this.coordinates;
+   
+    return {
+      x: xCurrent - xNext,
+      y: yCurrent - yNext,
+    };
   }
 }